updated the team page
